Finally actually fix GC crash in xns_read_list by initializing variables.
This heisenbug was very mysterious, randomly dissapearing and reappearing after
what seemed to be unrelated changes to vastly disparate parts of the code or
fixes to other bugs.  This bug was only triggered when a GC happened during
xns_read_list (cdr was uninitialized except for when reading in a dotted pair terminator).

Additionally, add documentation and change the type of xns_gc_register in an
attempt to prevent this from happening again, although only clang will actually
generate a warning about this.