Implement all non-numerical currently defined primitives.
Note that I have found that all gc-registered variables should be
made volatile -- passing a pointer to the local variable does NOT
let the compiler know that the garbage collector might modify it at
any time (see #1).  As a temporary testing fix, the default heap size has been
changed to 4 MB (this is closer to a production value, but a heap size
of 128 bytes helps testing).